Add a missing header and tidy up a few places
The buffer for precalculated dialog data was set to 2KBs from when I
tried to add the text from there, but that caused the URL autodetection
to not trigger when initializing the license dialog. There might be a
way to process the existing text after the fact, but I can't be bothered
to find out, since this would only save a single WM_SETTEXT message.

The <bit> header was not included for bit_cast.

wWinMain has SAL annotations in Windows.h and VS was saying the
signatures didn't match, so I just copied the signature from Windows.h.